pyproject.toml
[project]
name = "meno-core"
version = "0.2.0"
description = "Meno Core: The backbone of Meno's intelligent systems."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "apscheduler>=3.11.1",
    "fastapi>=0.121.3",
    "future>=1.0.0",
    #    "graspologic>=3.4.4",
    "lightrag-hku>=1.4.9.8",
    "nano-vectordb>=0.0.4.3",
    "nltk>=3.9.2",
    "numpy>=1.26.4,<2.0.0",
    "scipy>=1.15.1",
    "openai>=2.8.1",
    "pipmaster>=1.0.10",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pytz>=2025.2",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
    "rapidfuzz>=3.14.3",
    "rank-bm25>=0.2.2",
    "zvec>=0.2.0",
]

[project.scripts]
meno-api = "meno_core.api.main:run"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "build",
    "pre-commit",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F"]
ignore = [
    "E501",
]
